==> 4_all_analyses.Rmd <==
---
title: "Maladaptive plastic responses of flowering time to geothermal heating (Cerastium 2)"
subtitle: "Repeat and extend analyses done by Johan"
author : "Alicia Valdés"
output:
  pdf_document:
    latex_engine: xelatex
    toc: yes
    toc_depth: 4
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r load packages, include=FALSE}
library(tidyverse)
library(ggthemes)
library(knitr)
library(gridExtra)
library(DHARMa)
library(RColorBrewer)
library(broom)
library(ggpubr)
library(jtools)
library(kableExtra)
library(ggeffects)
library(MuMIn)
library(MASS)
library(segmented)
library(effects)
library(purrr)
library(lubridate)
library(ggforce)
library(lmtest)
library(car)
library(viridis)
library(ggrepel)
library(grid)
```

```{r Define ggplot themes and palettes, include=FALSE}
my_theme <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(legend.position="none")+theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
my_theme_legend <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
myPalette <- colorRampPalette(brewer.pal(11, "YlOrRd"))
```

```{r, include=FALSE}
set_summ_defaults(digits = 3)
```

```{r function to check for overdispersion, include=FALSE}
overdisp_fun <- function(model) {
    rdf <- df.residual(model)
    rp <- residuals(model,type="pearson")
    Pearson.chisq <- sum(rp^2)
    prat <- Pearson.chisq/rdf
    pval <- pchisq(Pearson.chisq, df=rdf, lower.tail=FALSE)
    c(chisq=Pearson.chisq,ratio=prat,rdf=rdf,p=pval)
}
```

```{r load previously saved objects}
load("output/BCIs_selection_2017.RData")
load("output/BCIs_selection_2018.RData")
load("output/BCIs_selection_2017_abs.RData")
load("output/BCIs_selection_2018_abs.RData")
```

# Read data from loggers

```{r include=FALSE}
logger_data<-read_csv("data/edited/logger_data.csv")
logger_data_pairs<-read_csv("data/edited/logger_data_pairs.csv")
```

# Data preparation

Load data, keep variables needed and merge

```{r Load data}
data_2017<-read.table("data/edited/Cerastium2017_AV.csv",header=T,sep=",",
                      dec=".") 
data_2018<-read.table("data/edited/Cerastium2018_AV.csv",header=T,sep=",",
                      dec=".") 
data_2017<-subset(data_2017,INCLUDE==1)[c(1:5,7)]
data_2018<-subset(data_2018,INCLUDE==1)[c(1:4,7,9)]
names(data_2017)<-c("id","id_original","temp","ffd","nfl","nseed")
names(data_2018)<-c("id","id_original","temp","ffd","nfl","nseed")
data_2017$year<-"2017"
data_2018$year<-"2018"
data_2017$ffd_std<-as.numeric(scale(data_2017$ffd))
data_2017$nfl_std<-as.numeric(scale(log(data_2017$nfl)))
# First log, then standardized
data_2017$nseed_rel<-with(data_2017,nseed/mean(nseed))
data_2018$ffd_std<-as.numeric(scale(data_2018$ffd))
data_2018$nfl_std<-as.numeric(scale(log(data_2018$nfl)))
# First log, then standardized
data_2018$nseed_rel<-with(data_2018,nseed/mean(nseed))
mydata<-rbind(data_2017,data_2018)
mydata$id<-as.factor(mydata$id)
mydata$year<-as.factor(mydata$year)
head(mydata)
```

```{r include=FALSE}
sum(is.na(mydata))
```

# Distributions

Histograms

```{r fig.height=4, fig.width=7}
pivot_longer(mydata,cols=3:6) %>% 
  ggplot(aes(value)) +
    facet_wrap(~ name, scales = "free") +
    geom_histogram(fill="white", color="black")
```

Histograms by year

```{r fig.height=4, fig.width=7}
pivot_longer(mydata,cols=3:6) %>% 
  ggplot(aes(x=value,fill=year,color=year)) +
    facet_wrap(~ name, scales = "free") +
    geom_histogram(alpha=0.5)
```

QQplots

```{r fig.height=5, fig.width=7}
ggqqplot(mydata[3:7], x = c("ffd", "nfl", "nseed", "temp"),
         combine = TRUE,scales="free")
```

QQplots by year

```{r fig.height=5.5, fig.width=7}
ggqqplot(mydata[3:7], x = c("ffd", "nfl", "nseed", "temp"),color="year",
         combine = TRUE,scales="free")
```

# 1. How useful soil temperature is as a cue for air temperature

Do correlations between soil and air temperature vary with soil temperature?

We test this using logger pairs.

After looking at plots from individual pairs of loggers, make a new variable pair_problem:
- 0 = no problem
- 1 = some problems (remove)

Removed logger pairs with problems: 63,64,65,66,67,68,69,71,72,73)

Most loggers end up on June 6th, so using data until June 5th.

Using pairs of loggers where above and belowground logger are at less than 2 m distance.

```{r}
logger_data_pairs<-logger_data_pairs%>%
  mutate(pair_problem=ifelse(pair%in%c(63:69,71:73),1,0))%>%
  filter(pair_problem==0)%>% # remove pairs with problems
  mutate(date=date(datetime))%>%
  filter(!is.na(date))%>% # remove records with no info on date
  filter(datetime<"2018-06-06")%>% # keep only data until June 5
  filter(dist<2) # keep logger pairs with < 2 m distance
```

```{r}
length(unique(logger_data_pairs$logger_nr))
length(unique(subset(logger_data_pairs,species=="cer")$logger_nr))
```

80 loggers (40 pairs), of which 34 loggers in Cerastium plants, and 46 loggers in other plants.

For each date and logger pair, calculate mean, max and min of air and soil temperature (from, respectively, the above and belowground logger). Then, calculate the correlation coefficient for air and soil temperatures over the period May or April-May-June. Finally, regress these correlation coefficients on mean soil temperature (from the belowground logger) **for the same period (April-May-June)**.

## May

Linear models testing the effect of soil temperature on correlations between soil and air temperature:

```{r message=FALSE, warning=FALSE}
(logger_data_pairs%>%
     mutate(month = month(datetime))%>% # new variable "month"
     filter(month==5)%>% # keep data from may
     group_by(date,pair,above_below)%>% 
     summarise(mean=mean(temp,na.rm=T),max=max(temp),min=min(temp),
               plot=first(plot))%>% 
   #calculate mean, max and min of air and soil temperature 
     pivot_wider(names_from="above_below",values_from=c("mean","max","min"))%>%
     group_by(pair)%>%
     summarise(corr_airsoil_mean=cor(mean_A,mean_B,use="pairwise.complete.obs"),
               corr_airsoil_max=cor(max_A,max_B,use="pairwise.complete.obs"),
               corr_airsoil_min=cor(min_A,min_B,use="pairwise.complete.obs"),
               plot=first(plot))%>%
   # Calculate correlations air-soil temperatures
     pivot_longer(cols=corr_airsoil_mean:corr_airsoil_min,
                  names_to="measure",values_to="corr")%>%
  left_join(logger_data_pairs%>%
                 mutate(month = month(datetime))%>%
                 filter(month==5)%>%
                 filter(above_below=="B")%>%
                 group_by(pair)%>%
              summarise(meansoiltemp=mean(temp),dist=first(dist))))%>%
  # calculate mean soil temperature for may
  group_by(measure)%>%
  do(fitcorr = tidy(lm(corr~meansoiltemp, data = .))) %>% 
  unnest(fitcorr)%>%
  kable(digits=5)
```

```{r}
data_corr<-(logger_data_pairs%>%
     mutate(month = month(datetime))%>% # new variable "month"
     filter(month==5)%>% # keep data from may
     group_by(date,pair,above_below)%>% 
     summarise(mean=mean(temp,na.rm=T),max=max(temp),min=min(temp),
               plot=first(plot))%>% 
   #calculate mean, max and min of air and soil temperature 
     pivot_wider(names_from="above_below",values_from=c("mean","max","min"))%>%
     group_by(pair)%>%
     summarise(corr_airsoil_mean=cor(mean_A,mean_B,use="pairwise.complete.obs"),
               corr_airsoil_max=cor(max_A,max_B,use="pairwise.complete.obs"),
               corr_airsoil_min=cor(min_A,min_B,use="pairwise.complete.obs"),
               plot=first(plot))%>%
   # Calculate correlations air-soil temperatures
     pivot_longer(cols=corr_airsoil_mean:corr_airsoil_min,
                  names_to="measure",values_to="corr")%>%
  left_join(logger_data_pairs%>%
                 mutate(month = month(datetime))%>%
                 filter(month==5)%>%
                 filter(above_below=="B")%>%
                 group_by(pair)%>%
              summarise(meansoiltemp=mean(temp),dist=first(dist))))
  # calculate mean soil temperature for may
```

```{r}
model_mean<-lm(corr~meansoiltemp, 
               data = subset(data_corr,measure=="corr_airsoil_mean"))
model_max<-lm(corr~meansoiltemp, 
               data = subset(data_corr,measure=="corr_airsoil_max"))
model_min<-lm(corr~meansoiltemp, 
               data = subset(data_corr,measure=="corr_airsoil_min"))
```

Predictions of correlations for minimum and maximum temperatures:

```{r}
ggpredict(model_mean,terms="meansoiltemp[minmax]") 
ggpredict(model_max,terms="meansoiltemp[minmax]") 
ggpredict(model_min,terms="meansoiltemp[minmax]") 
```

### Figure 2: Correlations soil-air temperature vs soil temperature

```{r fig.height=6, fig.width=6, message=FALSE, warning=FALSE}
fig2<-(logger_data_pairs%>%
          mutate(month = month(datetime))%>% # new variable "month"
          filter(month==5)%>% # keep data from may
          group_by(date,pair,above_below)%>% 
          summarise(mean=mean(temp,na.rm=T),max=max(temp),min=min(temp),
                    plot=first(plot))%>% 
          #calculate mean, max and min of air and soil temperature 
          pivot_wider(names_from="above_below",
                      values_from=c("mean","max","min"))%>%
          group_by(pair)%>%
          summarise(corr_airsoil_mean=cor(mean_A,mean_B,
                                          use="pairwise.complete.obs"),
                    corr_airsoil_max=cor(max_A,max_B,
                                         use="pairwise.complete.obs"),
                    corr_airsoil_min=cor(min_A,min_B,
                                         use="pairwise.complete.obs"),
                    plot=first(plot))%>%
          # Calculate correlations air-soil temperatures
          pivot_longer(cols=corr_airsoil_mean:corr_airsoil_min,
                       names_to="measure",values_to="corr")%>%
          left_join(logger_data_pairs%>%
                      mutate(month = month(datetime))%>%
                      filter(month==5)%>%
                      filter(above_below=="B")%>%
                      group_by(pair)%>%
                      summarise(meansoiltemp=mean(temp),dist=first(dist))))%>%
  # calculate mean soil temperature for may
  ggplot(.,aes(x=meansoiltemp,y=corr,color=measure,fill=measure))+
  geom_smooth(method="lm",size=1)+geom_point(size=2,alpha=0.5)+
  xlab("Mean soil temperature (ºC)")+
  ylab("Correlation coefficient\nsoil-air temperature")+
  my_theme()+
  geom_text_repel(data=. %>% filter(corr<0),aes(label=pair))
fig2
ggsave(filename="output/figures/fig2.tiff",plot=fig2,
       width=12,height=10,units="cm",dpi=300)
```

## April-May_june

Linear models testing the effect of soil temperature on correlations between soil and air temperature:

```{r message=FALSE, warning=FALSE}
(logger_data_pairs%>%
     mutate(month = month(datetime))%>% # new variable "month"
     filter(month==4|month==5|month==6)%>% # keep data from april-june
     group_by(date,pair,above_below)%>% 
     summarise(mean=mean(temp,na.rm=T),max=max(temp),min=min(temp),
               plot=first(plot))%>% 
   #calculate mean, max and min of air and soil temperature 
     pivot_wider(names_from="above_below",values_from=c("mean","max","min"))%>%
     group_by(pair)%>%
     summarise(corr_airsoil_mean=cor(mean_A,mean_B,use="pairwise.complete.obs"),
               corr_airsoil_max=cor(max_A,max_B,use="pairwise.complete.obs"),
               corr_airsoil_min=cor(min_A,min_B,use="pairwise.complete.obs"),
               plot=first(plot))%>%
   # Calculate correlations air-soil temperatures
     pivot_longer(cols=corr_airsoil_mean:corr_airsoil_min,
                  names_to="measure",values_to="corr")%>%
  left_join(logger_data_pairs%>%
                 mutate(month = month(datetime))%>%
                 filter(month==4|month==5|month==6)%>%
                 filter(above_below=="B")%>%
                 group_by(pair)%>%
              summarise(meansoiltemp=mean(temp),dist=first(dist))))%>%
  # calculate mean soil temperature for april-june
  group_by(measure)%>%
  do(fitcorr = tidy(lm(corr~meansoiltemp, data = .))) %>% 
  unnest(fitcorr)%>%
  kable(digits=5)
```

### Appendix S2: Correlations soil-air temperature vs soil temperature (April-June)

```{r fig.height=6, fig.width=6, message=FALSE, warning=FALSE}
AppS2<-(logger_data_pairs%>%
          mutate(month = month(datetime))%>% # new variable "month"
          filter(month==4|month==5|month==6)%>% # keep data from april-june
          group_by(date,pair,above_below)%>% 
          summarise(mean=mean(temp,na.rm=T),max=max(temp),min=min(temp),
                    plot=first(plot))%>% 
          #calculate mean, max and min of air and soil temperature 
          pivot_wider(names_from="above_below",
                      values_from=c("mean","max","min"))%>%
          group_by(pair)%>%
          summarise(corr_airsoil_mean=cor(mean_A,mean_B,
                                          use="pairwise.complete.obs"),
                    corr_airsoil_max=cor(max_A,max_B,
                                         use="pairwise.complete.obs"),
                    corr_airsoil_min=cor(min_A,min_B,
                                         use="pairwise.complete.obs"),
                    plot=first(plot))%>%
          # Calculate correlations air-soil temperatures
          pivot_longer(cols=corr_airsoil_mean:corr_airsoil_min,
                       names_to="measure",values_to="corr")%>%
          left_join(logger_data_pairs%>%
                      mutate(month = month(datetime))%>%
                      filter(month==4|month==5|month==6)%>%
                      filter(above_below=="B")%>%
                      group_by(pair)%>%
                      summarise(meansoiltemp=mean(temp),dist=first(dist))))%>%
  # calculate mean soil temperature for april-june
  ggplot(.,aes(x=meansoiltemp,y=corr,color=measure,fill=measure))+
  geom_smooth(method="lm",size=1)+geom_point(size=2,alpha=0.5)+
  xlab("Mean soil temperature (ºC)")+
  ylab("Correlation coefficient\nsoil-air temperature")+
  my_theme()+
  geom_text_repel(data=. %>% filter(corr<0),aes(label=pair))
AppS2
ggsave(filename="output/figures/AppS2.tiff",plot=AppS2,
       width=12,height=10,units="cm",dpi=300)
```

# 2. Effect of temperature on FFD

Not including number of flowers as covariate, including quadratic effects of ffd.

```{r echo=TRUE}
FFD_2017_1<-lm(ffd~temp+I(temp^2),subset(mydata,year==2017))
summ(FFD_2017_1,vif=T,scale=T) # scale=T reports standardized coefs
FFD_2018_1<-lm(ffd~temp+I(temp^2),subset(mydata,year==2018))
summ(FFD_2018_1,vif=T,scale=T) # scale=T reports standardized coefs
```

Quadratic term of ffd significant in 2017 but not in 2018. Refit model for 2018 withouth quadratic term of ffd.

```{r echo=TRUE}
FFD_2018_2<-lm(ffd~temp,subset(mydata,year==2018))
summ(FFD_2018_2,scale=T)
```

Predictions of ffd for minimum and maximum temperatures:

```{r}
ggpredict(FFD_2017_1,terms="temp[minmax]")
# 180.75-173.63=8 days earlier on warmer soils
ggpredict(FFD_2018_2,terms="temp[minmax]") 
# 191.36-158.74=33 days earlier on warmer soils
```

## Model diagnostics

```{r}
simulationOutput_FFD_2017_1<-simulateResiduals(fittedModel=FFD_2017_1,
                                               n=5000)
simulationOutput_FFD_2018_1<-simulateResiduals(fittedModel=FFD_2018_1,
                                               n=5000)
smulationOutput_FFD_2018_2<-simulateResiduals(fittedModel=FFD_2018_2,
                                               n=5000)
```

FFD_2017_1

```{r fig.height=4, fig.width=7}
plot(simulationOutput_FFD_2017_1)
```

Residuals against temp

```{r fig.height=4, fig.width=7}
plotResiduals(simulationOutput_FFD_2017_1,
     form = subset(mydata,year==2017)$temp)
```

2018:

```{r fig.height=4, fig.width=7}
plot(simulationOutput_FFD_2018_1)
```

Residuals against temp

```{r fig.height=4, fig.width=7}
plotResiduals(simulationOutput_FFD_2018_1,
     form = (subset(mydata,year==2018)$temp)^2)
```

Models should be OK.

## Piecewise regresssion (not shown).

```{r echo=TRUE}
FFD_2017_segm<-segmented(lm(ffd~temp,subset(mydata,year==2017)),
                         seg.Z=~temp,psi=20)
slope(FFD_2017_segm)
AIC(FFD_2017_1,FFD_2017_segm)
```

CIs for the slope of the second segment include zero (means it is not significantly different from zero?). Very little difference in AIC (\<2), so the piecewise regression is not better.

```{r echo=TRUE}
FFD_2018_segm<-segmented(lm(ffd~temp,subset(mydata,year==2018)),
                         seg.Z=~temp,psi=15)
slope(FFD_2018_segm)
AIC(FFD_2018_1,FFD_2018_segm)
```

CIs for the slope of the second segment include zero (means it is not significantly different from zero?). Difference in AIC \>2 so the piecewise regression seems to be better.

### Plot 2018 (not used)

```{r fig.height=4, fig.width=4}
# ggplot(subset(mydata,year==2018),aes(x=temp,y=ffd))+my_theme()+
#   geom_ribbon(data=predict_FFD_2018_1,
#               aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
#                 fill="#00BFC4",alpha=0.5)+
#   geom_line(data=predict_FFD_2018_1,
#             aes(x=x,y=predicted),size=1,color="#00BFC4")+
#   geom_ribbon(data=subset(mydata,year==2018),
#               aes(x=temp,y=broken.line(FFD_2018_segm)$fit,
#                   ymin=broken.line(FFD_2018_segm)$fit-
#                     broken.line(FFD_2018_segm)$se.fit,
#                   ymax=broken.line(FFD_2018_segm)$fit+
#                     broken.line(FFD_2018_segm)$se.fit),
#               fill="grey",alpha=0.5)+
#   geom_line(data=subset(mydata,year==2018),
#             aes(x=temp,y=broken.line(FFD_2018_segm)$fit),
#             size=1,color="darkgrey")+
#   geom_point()+xlab("Temperature (ºC)")+ylab("FFD")
```

Keep the linear model, as the piecewise one does not seem to fit much better (the reduction in AIC was also small).

# 3. Effect of temperature on fitness

```{r echo=TRUE}
fitness_2017_1<-lm(nseed~temp+I(temp^2)+log(nfl),subset(mydata,year==2017))
summ(fitness_2017_1,vif=T,scale=T) # scale=T reports standardized coefs
fitness_2018_1<-lm(nseed~temp+I(temp^2)+log(nfl),subset(mydata,year==2018))
summ(fitness_2018_1,vif=T,scale=T) # scale=T reports standardized coefs
```

Quadratic terms of ffd not significant. Refit models withouth quadratic terms of ffd.

```{r echo=TRUE}
fitness_2017_2<-lm(nseed~temp+log(nfl),subset(mydata,year==2017))
summ(fitness_2017_2,vif=T,scale=T) # scale=T reports standardized coefs
fitness_2018_2<-lm(nseed~temp+log(nfl),subset(mydata,year==2018))
summ(fitness_2018_2,vif=T,scale=T) # scale=T reports standardized coefs
```

## Model diagnostics

```{r}
simulationOutput_fitness_2017_2<-simulateResiduals(fittedModel=fitness_2017_2,
                                                   n=5000)
simulationOutput_fitness_2018_2<-simulateResiduals(fittedModel=fitness_2018_2,
                                                   n=5000)
```

qq-plot and plot of residuals vs. predicted:

2017:

```{r fig.height=4, fig.width=7}
plot(simulationOutput_fitness_2017_2)
```

2018:

```{r fig.height=4, fig.width=7}
plot(simulationOutput_fitness_2018_2)
```

Quite bad looking! We should try another distribution.

## GLMs with poisson distribution

```{r echo=TRUE}
mydata$n_seed_round<-with(mydata,round(nseed))
fitness_2017_3<-glm(n_seed_round~temp+I(temp^2)+log(nfl),
                    subset(mydata,year==2017),family="poisson")
summ(fitness_2017_3,vif=T,scale=T)
fitness_2018_3<-glm(n_seed_round~temp+I(temp^2)+log(nfl),
                    subset(mydata,year==2018),family="poisson")
summ(fitness_2018_3,vif=T,scale=T)
```

```{r echo=TRUE}
overdisp_fun(fitness_2017_3) 
overdisp_fun(fitness_2018_3)
```

There is significant overdispersion.

## GLMs with negative binomial distribution

```{r echo=TRUE}
fitness_2017_4<-glm.nb(n_seed_round~temp+I(temp^2)+log(nfl),
                       subset(mydata,year==2017))
summ(fitness_2017_4,vif=T,scale=T) # scale=T reports standardized coefs
fitness_2018_4<-glm.nb(n_seed_round~temp+I(temp^2)+log(nfl),
                       subset(mydata,year==2018))
summ(fitness_2018_4,vif=T,scale=T) # scale=T reports standardized coefs
```

Quadratic terms of ffd not significant. Refit models withouth quadratic terms of ffd.

```{r echo=TRUE}
fitness_2017_5<-glm.nb(n_seed_round~temp+log(nfl),subset(mydata,year==2017))
summ(fitness_2017_5,vif=T,scale=T)
fitness_2018_5<-glm.nb(n_seed_round~temp+log(nfl),subset(mydata,year==2018))
summ(fitness_2018_5,vif=T,scale=T)
```

Predictions of fitness for minimum and maximum temperatures:

```{r}
ggpredict(fitness_2017_5,terms="temp[minmax]")
ggpredict(fitness_2018_5,terms="temp[minmax]") 
```

### Model diagnostics

```{r}
simulationOutput_fitness_2017_5<-simulateResiduals(fittedModel=fitness_2017_5,
                                                   n=5000)
simulationOutput_fitness_2018_5<-simulateResiduals(fittedModel=fitness_2018_5,
                                                   n=5000)
```

qq-plot and plot of residuals vs. predicted:

2017:

```{r echo=FALSE, fig.height=4, fig.width=7}
plot(simulationOutput_fitness_2017_5)
```

2018:

```{r echo=FALSE, fig.height=4, fig.width=7}
plot(simulationOutput_fitness_2018_5)
```

Some problems but maybe not so bad. Need to look a bit more into this later.

# Figure 3: Effects of temperature on ffd, size and fitness in 2018

Model predictions ffd : based on models FFD_2017_1 (with quadratic term of ffd) and FFD_2018_2 (without quadratic term of ffd).

```{r}
predict_FFD_2017<-ggpredict(FFD_2017_1,terms = "temp [all]")
predict_FFD_2018<-ggpredict(FFD_2018_2,terms = "temp [all]")
```

Model predictions fitness : based on models fitness_2017_5 and fitness_2018_5.

```{r}
predict_fitness_2017<-ggpredict(fitness_2017_5,terms = "temp [all]")
predict_fitness_2018<-ggpredict(fitness_2018_5,terms = "temp [all]")
```


```{r fig.height=4, fig.width=8, message=FALSE, warning=FALSE}
fig3<-
  grid.arrange(
    # ffd
    ggplot(subset(mydata,year==2018),aes(x=temp,y=ffd))+
      xlab("Soil temperature (ºC)")+ylab("FFD (Day of year)")+my_theme()+
      geom_ribbon(data=predict_FFD_2018,
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                  fill="grey",alpha=0.7)+
      geom_line(data=predict_FFD_2018,
                aes(x=x,y=predicted),size=1,color="black")+
      geom_point(size=2,alpha=0.2)+
      ggtitle("A)")+theme(plot.title=element_text(hjust=-0.20,vjust=-3))+
      theme(plot.margin = unit(c(-0.6,0.3,0,0.3), "cm")),
    # fitness
    ggplot(subset(mydata,year==2018),aes(x=temp,y=nseed))+
      xlab("Soil temperature (ºC)")+ylab("Fitness (number of seeds)")+
      my_theme()+
      geom_ribbon(data=predict_fitness_2018,
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                  fill="grey",alpha=0.7)+
      geom_line(data=predict_fitness_2018,
                aes(x=x,y=predicted),size=1,color="black")+
      geom_point(size=2,alpha=0.2)+
      ggtitle("B)")+theme(plot.title=element_text(hjust=-0.20,vjust=-3))+
      theme(plot.margin = unit(c(-0.6,0.3,0,0.3), "cm")),
    ncol=2)
ggsave(filename="output/figures/fig3.tiff",plot=fig3,
       width=20,height=8,units="cm",dpi=300)
```

# Figure S1: Effects of temperature on ffd, size and fitness in 2017

```{r fig.height=4, fig.width=8, message=FALSE, warning=FALSE}
figS1<-
  grid.arrange(
    # ffd
    ggplot(subset(mydata,year==2017),aes(x=temp,y=ffd))+
      xlab("Soil temperature (ºC)")+ylab("FFD (Julian date)")+my_theme()+
      geom_ribbon(data=predict_FFD_2017,
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                  fill="grey",alpha=0.7)+
      geom_line(data=predict_FFD_2017,
                aes(x=x,y=predicted),size=1,color="black")+
      geom_point(size=2,alpha=0.2)+
      ggtitle("A)")+theme(plot.title=element_text(hjust=-0.24,vjust=-3))+
      theme(plot.margin = unit(c(-0.6,0.3,0,0.3), "cm")),
    # fitness
    ggplot(subset(mydata,year==2017),aes(x=temp,y=nseed))+
      xlab("Soil temperature (ºC)")+ylab("Fitness (number of seeds)")+
      my_theme()+
      geom_ribbon(data=predict_fitness_2017,
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                  fill="grey",alpha=0.7)+
      geom_line(data=predict_fitness_2017,
                aes(x=x,y=predicted),size=1,color="black")+
      geom_point(size=2,alpha=0.2)+
      ggtitle("B)")+theme(plot.title=element_text(hjust=-0.20,vjust=-3))+
      theme(plot.margin = unit(c(-0.6,0.3,0,0.3), "cm")),
    ncol=2)
ggsave(filename="output/figures/figS1.tiff",plot=figS1,
       width=20,height=8,units="cm",dpi=300)
```

# Possible Appendix: Are instantaneous measures of soil temperature representative for the conditions during the entire spring/growing season?

Correlations logger temperature - instant temperature

## April-June

For each logger\_nr, get mean temperature during April-June and compare with temp\_term (which was measured with a thermometer at 10 cm depth on May 2017):

```{r message=FALSE, warning=FALSE, fig.height=4, fig.width=5}
alltemps<-logger_data%>%
  mutate(month = month(datetime)) %>%
  filter(month==4|month==5|month==6)%>%
  filter(above_below=="B")%>%
  mutate(date=date(datetime))%>%
  filter(!is.na(date))%>% # remove records with no info on date
  filter(datetime<"2018-06-06")%>% # keep only data until June 5
  group_by(logger_nr) %>%
  summarize(mean_logger=mean(temp),temp_term=mean(temp_term))
ggplot(alltemps,aes(x=temp_term,y=mean_logger))+
  geom_smooth(method="lm",size=1,color="black")+geom_point(size=2,alpha=0.2)+
  geom_abline(lty="dashed",alpha=0.8)+
  xlab("Instant soil temperature measured in May")+
  ylab("Mean temperature\nfrom logger data (April-June)")+my_theme()
summ(lm(mean_logger~temp_term,alltemps))
```

Correlation mean temperature from logger data (only belowground loggers) and soil temperature measured in may with thermometer:

```{r}
with(alltemps,cor(mean_logger,temp_term))
```

# 4. Effect of temperature on selection on FFD

```{r echo=TRUE}
selection_2017_1<-lm(nseed_rel~ffd_std*temp+ffd_std*I(temp^2)+nfl_std,
                     subset(mydata,year==2017))
selection_2018_1<-lm(nseed_rel~ffd_std*temp+ffd_std*I(temp^2)+nfl_std,
                     subset(mydata,year==2018))
summ(selection_2017_1,scale=T)
summ(selection_2018_1,scale=T)
```

Quadratic terms of ffd not significant. Refit models withouth quadratic terms of ffd.

```{r echo=TRUE}
selection_2017_2<-lm(nseed_rel~ffd_std*temp+nfl_std,
                     subset(mydata,year==2017))
selection_2018_2<-lm(nseed_rel~ffd_std*temp+nfl_std,
                     subset(mydata,year==2018))
summ(selection_2017_2,scale=T,digits=5)
summ(selection_2018_2,scale=T,digits=5)
```

## Model diagnostics

```{r}
simulationOutput_selection_2017_2<-simulateResiduals(
  fittedModel=selection_2017_2,n=5000)
simulationOutput_selection_2018_2<-simulateResiduals(
  fittedModel=selection_2018_2,n=5000)
```

qq-plot and plot of residuals vs. predicted:

2017:

```{r fig.height=4, fig.width=7}
plot(simulationOutput_selection_2017_2)
```

2018:

```{r fig.height=4, fig.width=7}
plot(simulationOutput_selection_2018_2)
```

Quite bad looking! If we want to keep the linear model with normal distribution (i.e. a "classic" selection model), we can assess significances using BCa intervals.

## BCa intervals

### 2017

```{r eval=FALSE, include=FALSE}
# ffd
slp <- function(selection_2017_2) coef(selection_2017_2)[2]
b <- car::Boot(selection_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_ci_17 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# temp
slp <- function(selection_2017_2) coef(selection_2017_2)[3]
b <- car::Boot(selection_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
temp_ci_17 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# nfl
slp <- function(selection_2017_2) coef(selection_2017_2)[4]
b <- car::Boot(selection_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
nfl_ci_17 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# ffd:temp
slp <- function(selection_2017_2) coef(selection_2017_2)[5]
b <- car::Boot(selection_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_temp_ci_17 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_selection_2017 <- cbind(
  rbind(ffd_ci_17[1,] ,temp_ci_17[1,], nfl_ci_17[1,], ffd_temp_ci_17[1,]),
  rbind(ffd_ci_17[2,] ,temp_ci_17[2,], nfl_ci_17[2,], ffd_temp_ci_17[2,])
)
colnames(BCIs_selection_2017)<-c("lower","upper")
rownames(BCIs_selection_2017) <- c("ffd","temp","nfl","ffd:temp")
save(BCIs_selection_2017,file="output/BCIs_selection_2017.RData")
```

```{r}
BCIs_selection_2017
```

FFD significant according to BCa intervals.

### 2018

```{r eval=FALSE, include=FALSE}
# ffd
slp <- function(selection_2018_2) coef(selection_2018_2)[2]
b <- car::Boot(selection_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_ci_18 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# temp
slp <- function(selection_2018_2) coef(selection_2018_2)[3]
b <- car::Boot(selection_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
temp_ci_18 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# nfl
slp <- function(selection_2018_2) coef(selection_2018_2)[4]
b <- car::Boot(selection_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
nfl_ci_18 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# ffd:temp
slp <- function(selection_2018_2) coef(selection_2018_2)[5]
b <- car::Boot(selection_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_temp_ci_18 <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_selection_2018 <- cbind(
  rbind(ffd_ci_18[1,] ,temp_ci_18[1,], nfl_ci_18[1,], ffd_temp_ci_18[1,]),
  rbind(ffd_ci_18[2,] ,temp_ci_18[2,], nfl_ci_18[2,], ffd_temp_ci_18[2,])
)
colnames(BCIs_selection_2018)<-c("lower","upper")
rownames(BCIs_selection_2018) <- c("ffd","temp","nfl","ffd:temp")
save(BCIs_selection_2018,file="output/BCIs_selection_2018.RData")
```

```{r}
BCIs_selection_2018
```

The significances according to the BCa intervals are similar to the ones given in the model summary.

# Figure 4: Effects of temperature on selection in 2018

```{r}
pred_fitness<-ggpredict(selection_2018_2,
                        terms = c("ffd_std [all]","temp [all]"))
```

```{r fig.height=3.5, fig.width=6}
ggplot(pred_fitness,aes(x,predicted,colour=group,fill=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.5)+my_theme()+
  scale_color_viridis()+
  theme(legend.position="right")+labs(colour="Soil temperature (ºC)")+
  xlab("Standardized FFD")+ylab("Relative fitness")
ggsave(filename="output/figures/fig4.tiff",width=14,height=8,units="cm",dpi=300)
```

# Possble Appendix: Relationships between std FFD and relative fitness for different parts of the distribution of soil temperature

```{r}
summary(subset(mydata,year==2018)$temp) # Get quartile values
mean(subset(mydata,year==2018&temp<=6.875)$temp) 
# Mean cat 1 = 5.784615
mean(subset(mydata,year==2018&temp>6.875&temp<=9.3)$temp) 
# Mean cat 2 = 7.95
mean(subset(mydata,year==2018&temp>9.3&temp<=17.075)$temp)
# Mean cat 3 = 12.45
mean(subset(mydata,year==2018&temp>17.075)$temp) 
# Mean cat 4 = 23.28846

pred_fitness_cats<-rbind(
  (data.frame(ggpredict(selection_2018_2,
                        terms = c("ffd_std [all]","temp [5.784615]")))%>%
     mutate(temp_cat=1)%>% 
     rename(ffd_std=x, fitness=predicted,temp=group)),
  (data.frame(ggpredict(selection_2018_2,
                        terms = c("ffd_std [all]","temp [7.95]")))%>%
     mutate(temp_cat=2)%>%
     rename(ffd_std=x, fitness=predicted,temp=group)),
  (data.frame(ggpredict(selection_2018_2,
                        terms = c("ffd_std [all]","temp [12.45]")))%>%
     mutate(temp_cat=3)%>%
     rename(ffd_std=x, fitness=predicted,temp=group)),
  (data.frame(ggpredict(selection_2018_2,
                        terms = c("ffd_std [all]","temp [23.28846]")))%>%
     mutate(temp_cat=4)%>%
     rename(ffd_std=x, fitness=predicted,temp=group)))

label_names1 <- list(
  '1'="First quarter\nMean temperature = 5.8 ºC",
  '2'="Second quarter\nMean temperature = 8.0 ºC",
  '3'="Third quarter\nMean temperature = 12.5 ºC",
  '4'="Fourth quarter\nMean temperature = 23.3 ºC"
)

labeller_function1 <- function(variable,value){
  return(label_names1[value])
}

```

```{r}
ggplot(subset(mydata,year==2018)%>%
         # Define 4 temp categories based on quartiles
         mutate(temp_cat=as.factor(
           ifelse(temp<=6.875,1,
                  ifelse(temp>6.875&temp<=9.3,2,
                         ifelse(temp>9.3&temp<=17.075,3,4))))),
       aes(x=ffd_std,y=nseed_rel))+
  facet_grid(~temp_cat,scales="free",
             labeller=labeller(temp_cat=labeller_function1))+
  geom_jitter(size=1.5,alpha=0.3,width=0.05)+
  geom_line(data=pred_fitness_cats,
            aes(x=ffd_std,y=fitness,color=temp_cat),size=1)+
  geom_ribbon(data=pred_fitness_cats,aes(x=ffd_std,y=fitness,
                                     ymin=conf.low,ymax=conf.high,
                                     fill=temp_cat),alpha=0.3)+
  my_theme()+scale_color_viridis(labels=NULL)+scale_fill_viridis(labels=NULL)+
  theme(legend.position="top")+labs(colour="Soil temperature (ºC)")+
  xlab("Standardized FFD")+
  ylab("Relative fitness")
  theme(strip.text.x=element_text(margin=margin(2,0,2,0)))+
  guides(fill=FALSE)
#ggsave()
# Fix: two legends appearing in figure
```

# Figure S2: Effects of ffd on relative fitness in 2018

```{r}
pred_fitness_18<-ggpredict(selection_2017_2,
                        terms = c("ffd_std [all]"))
```

```{r fig.height=3.5, fig.width=4}
ggplot(subset(mydata,year==2018),aes(x=ffd_std,y=nseed_rel))+
  xlab("Standardized FFD")+ylab("Relative fitness")+my_theme()+
  geom_ribbon(data=pred_fitness_18,
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="grey",alpha=0.7)+
  geom_line(data=pred_fitness_18,aes(x=x,y=predicted),size=1,color="black")+
  geom_point(size=2,alpha=0.2)
ggsave(filename="output/figures/figS2.tiff",width=10,height=8,
       units="cm",dpi=300)
```

# 5. Effect of temperature on the relationship absolute fitness-FFD

```{r echo=TRUE}
selectionabs_2017_1<-lm(nseed~ffd*temp+ffd*I(temp^2)+log(nfl),
                     subset(mydata,year==2017))
selectionabs_2018_1<-lm(nseed~ffd*temp+ffd*I(temp^2)+log(nfl),
                     subset(mydata,year==2018))
summ(selectionabs_2017_1,scale=F)
summ(selectionabs_2018_1,scale=F)
```

Quadratic terms of ffd not significant. Refit models withouth quadratic terms of ffd.

```{r echo=TRUE}
selectionabs_2017_2<-lm(nseed~ffd*temp+log(nfl),subset(mydata,year==2017))
selectionabs_2018_2<-lm(nseed~ffd*temp+log(nfl),subset(mydata,year==2018))
summ(selectionabs_2017_2,scale=F)
summ(selectionabs_2018_2,scale=F)
```

## Model diagnostics

```{r}
simulationOutput_selectionabs_2017_2<-simulateResiduals(
  fittedModel=selectionabs_2017_2,n=5000)
simulationOutput_selectionabs_2018_2<-simulateResiduals(
  fittedModel=selectionabs_2018_2,n=5000)
```

qq-plot and plot of residuals vs. predicted:

2017:

```{r fig.height=4, fig.width=7}
plot(simulationOutput_selectionabs_2017_2)
```

2018:

```{r fig.height=4, fig.width=7}
plot(simulationOutput_selectionabs_2018_2)
```

Quite bad looking! If we want to keep the linear model with normal distribution (i.e. a "classic" selection model), we can assess significances using BCa intervals.

## BCa intervals

### 2017

```{r eval=FALSE, include=FALSE}
# ffd
slp <- function(selectionabs_2017_2) coef(selectionabs_2017_2)[2]
b <- car::Boot(selectionabs_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_ci_17_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# temp
slp <- function(selectionabs_2017_2) coef(selectionabs_2017_2)[3]
b <- car::Boot(selectionabs_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
temp_ci_17_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# nfl
slp <- function(selectionabs_2017_2) coef(selectionabs_2017_2)[4]
b <- car::Boot(selectionabs_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
nfl_ci_17_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# ffd:temp
slp <- function(selectionabs_2017_2) coef(selectionabs_2017_2)[5]
b <- car::Boot(selectionabs_2017_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_temp_ci_17_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_selection_2017_abs <- cbind(
  rbind(ffd_ci_17_abs[1,] ,temp_ci_17_abs[1,], nfl_ci_17_abs[1,],
        ffd_temp_ci_17_abs[1,]),
  rbind(ffd_ci_17_abs[2,] ,temp_ci_17_abs[2,], nfl_ci_17_abs[2,],
        ffd_temp_ci_17_abs[2,])
)
colnames(BCIs_selection_2017_abs)<-c("lower","upper")
rownames(BCIs_selection_2017_abs) <- c("ffd","temp","nfl","ffd:temp")
save(BCIs_selection_2017_abs,file="output/BCIs_selection_2017_abs.RData")
```

```{r}
BCIs_selection_2017_abs
```

FFD significant according to BCa intervals.

### 2018

```{r eval=FALSE, include=FALSE}
# ffd
slp <- function(selectionabs_2018_2) coef(selectionabs_2018_2)[2]
b <- car::Boot(selectionabs_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_ci_18_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# temp
slp <- function(selectionabs_2018_2) coef(selectionabs_2018_2)[3]
b <- car::Boot(selectionabs_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
temp_ci_18_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# nfl
slp <- function(selectionabs_2018_2) coef(selectionabs_2018_2)[4]
b <- car::Boot(selectionabs_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
nfl_ci_18_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# ffd:temp
slp <- function(selectionabs_2018_2) coef(selectionabs_2018_2)[5]
b <- car::Boot(selectionabs_2018_2,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ffd_temp_ci_18_abs <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)
```

```{r eval=FALSE, include=FALSE}
# Save confidence intervals as a table
BCIs_selection_2018_abs <- cbind(
  rbind(ffd_ci_18_abs[1,] ,temp_ci_18_abs[1,], nfl_ci_18_abs[1,],
        ffd_temp_ci_18_abs[1,]),
  rbind(ffd_ci_18_abs[2,] ,temp_ci_18_abs[2,], nfl_ci_18_abs[2,],
        ffd_temp_ci_18_abs[2,])
)
colnames(BCIs_selection_2018_abs)<-c("lower","upper")
rownames(BCIs_selection_2018_abs) <- c("ffd","temp","nfl","ffd:temp")
save(BCIs_selection_2018_abs,file="output/BCIs_selection_2018_abs.RData")
```

```{r}
BCIs_selection_2018_abs
```

The significances according to the BCa intervals are similar to the ones given in the model summary.

```{r include=FALSE}
sessionInfo()
```
